airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
psycopg2-binary==2.9.6  # This is the PostgreSQL adapter for Python
requests==2.28.2  # This is for making HTTP requests to the World Bank API